Adds FormBuilderFormBase as another layer between shared FormBuilderBase and forms like FormInfo that deal with the top level of a Digital Form.
